% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create base grid for TICA}
\usage{
create_grid(aoi, crs, resolution, full_cells = TRUE, shape = "hex")
}
\arguments{
\item{aoi}{Path to file of area of interest.}

\item{crs}{Coordinate reference system}

\item{resolution}{Set the resolution of the grid in degrees.}

\item{full_cells}{Choose if cells should completely inside the biomes.}

\item{shape}{Shape of the grid cells.}
}
\value{
A tibble.
}
\description{
Create base grid for TICA
}
