% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visualizations.R
\name{create_visualizations}
\alias{create_visualizations}
\alias{eda_histogram}
\alias{eda_cumulative_distribution}
\alias{eda_spatial_distribution}
\alias{eda_time_series}
\alias{eda_colsum}
\alias{eda_summary_table}
\title{Title Create data visualizations}
\usage{
create_visualizations(
  f,
  data,
  group_facet = FALSE,
  group_variable = NULL,
  viz_title = NULL,
  x_title = NULL,
  y_title = NULL,
  out_format = c("png", "rdata"),
  out_filename = NULL,
  out_path = NULL,
  out_width = 15,
  out_height = 11,
  ...
)

eda_histogram(data, variable, n_bins, scale_transform, x_lim, ...)

eda_cumulative_distribution(
  data,
  variable,
  quantiles_list,
  scale_transform,
  ...
)

eda_spatial_distribution(data, variable, variable_label, base_map, ...)

eda_time_series(data, variable, ts_type = c("step", "bar", "line"), ...)

eda_colsum(data, variable, cat_variable, scale_transform, ...)

eda_summary_table(
  data,
  variable,
  rowname_variable,
  viz_title,
  out_filename = NULL,
  out_path = NULL
)
}
\arguments{
\item{f}{Type of visualization}

\item{data}{A tibble to provide data}

\item{group_facet}{Toggle faceting by group}

\item{group_variable}{Set variable to create facet label}

\item{viz_title}{Visualization title}

\item{x_title}{Horizontal axis title}

\item{y_title}{Vertical axis title}

\item{out_format}{The format of the visualization output}

\item{out_filename}{Name of the file}

\item{out_path}{Path of the file}

\item{out_width}{Width in centimeters of the .png file}

\item{out_height}{Height in centimeters of the .png file}

\item{...}{Parameters passed to visualization functions}

\item{variable}{The variable to be represented in the visualization}

\item{n_bins}{Number of bins of histogram}

\item{scale_transform}{A transform for the horizontal axis}

\item{x_lim}{Limit of x axis}

\item{quantiles_list}{List of percentiles to be calculated}

\item{variable_label}{Set the labels of the table variables}

\item{base_map}{An sf object to serve as base map}

\item{ts_type}{The type of the time series visualization}

\item{cat_variable}{Set the catgorical variable}

\item{rowname_variable}{Set the variable used as table row name}
}
\value{
A ggplot
}
\description{
Title Create data visualizations
}
\examples{
\dontrun{
  create_visualizations()
}

}
