% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_external_data.R
\name{download_external_data}
\alias{download_external_data}
\title{Download external data}
\usage{
download_external_data(
  dest_dir = "./data/external/",
  timespan = 1985:2021,
  aoi = TRUE,
  lulc = TRUE
)
}
\arguments{
\item{dest_dir}{Destination directory for external files}

\item{timespan}{Set the years to download data}

\item{aoi}{Download area of interest}

\item{lulc}{Download land use data}
}
\value{
A tibble
}
\description{
Download external data
}
\examples{
\dontrun{
  download_external_data(lulc = TRUE)
}

}
