% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_external_data.R
\name{preprocess_external_data}
\alias{preprocess_external_data}
\title{Process external data to base grid}
\usage{
preprocess_external_data(
  external_data = "conservation_units",
  preprocess_function
)
}
\arguments{
\item{external_data}{a}

\item{preprocess_function}{a}
}
\value{
An sf object
}
\description{
Process external data to base grid
}
