% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_conservation_units.R
\name{process_conservation_units}
\alias{process_conservation_units}
\title{Process conservation units data to base grid}
\usage{
process_conservation_units(base_grid_path, timespan)
}
\arguments{
\item{base_grid_path}{Path to the base grid file}

\item{timespan}{Set the years for the data in the base grid}
}
\value{
A tibble
}
\description{
Process conservation units data to base grid
}
\examples{
\dontrun{
  uc_grid <- process_conservation_units("./inst/base_grid.fgb", 1985:2021)
}

}
