% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_external_data.R
\name{process_external_data}
\alias{process_external_data}
\alias{process_conservation_units}
\alias{process_indigenous_lands}
\alias{process_land_use}
\alias{process_federal_roads}
\title{Process external data to base grid}
\usage{
process_external_data()

process_conservation_units(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/uc/uc.fgb",
  timespan = 1985:2021
)

process_indigenous_lands(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/il/il.fgb",
  timespan = 1985:2021
)

process_land_use(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/lulc/",
  timespan = 1985:2021,
  natural_class = c(3, 4, 11, 12),
  pasture_class = c(15),
  temporary_crop_class = c(39, 20, 40, 62, 41),
  perennial_crop_class = c(46, 47, 48),
  forest_plantation_class = c(9),
  mosaic_class = c(21)
)

process_federal_roads(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/fr/fr.fgb",
  timespan = 1985:2021
)
}
\arguments{
\item{base_grid_path}{Path to the base grid file}

\item{external_data_path}{Set path to external data file}

\item{timespan}{Set the years for the data in the base grid}

\item{natural_class}{Set classes of natural cover}

\item{pasture_class}{Set classes for pasture cover}

\item{temporary_crop_class}{Set classes for temporary crops cover}

\item{perennial_crop_class}{Set classes for perennial crops cover}

\item{forest_plantation_class}{Set classes for forest plantation cover}

\item{mosaic_class}{Set classes for mixed cover}
}
\value{
A tibble
}
\description{
Process external data to base grid
}
\examples{
\dontrun{
  uc_grid <- process_conservation_units("./inst/base_grid.fgb", 1985:2021)
}

}
