% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model_ensemble}
\alias{run_model_ensemble}
\alias{run_model}
\title{Sample Gridded data}
\usage{
run_model_ensemble(filled_data, sampled_data)

run_model(filled_data, sampled_data)
}
\arguments{
\item{filled_data}{a}
}
\value{
A tibble
}
\description{
Sample Gridded data
}
