% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample Gridded data}
\usage{
sample_data(filled_data, year_min, year_max, seed)
}
\arguments{
\item{filled_data}{a}
}
\value{
A tibble
}
\description{
Sample Gridded data
}
