% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create base grid for TICA}
\usage{
create_grid(resolution, full_cells)
}
\arguments{
\item{resolution}{Set the resolution of the grid in degrees.}

\item{full_cells}{Choose if cells should completely inside the biomes.}
}
\value{
A tibble.
}
\description{
Create base grid for TICA
}
\examples{
base_grid <- create_grid(resolution = 0.2, full_cells = TRUE)
}
