% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create base grid for TICA}
\usage{
create_grid(
  aoi_path = "./data/external/aoi/aoi.fgb",
  resolution = 0.5,
  full_cells = TRUE,
  shape = "hex"
)
}
\arguments{
\item{aoi_path}{Path to file of area of interest.}

\item{resolution}{Set the resolution of the grid in degrees.}

\item{full_cells}{Choose if cells should completely inside the biomes.}

\item{shape}{Shape of the grid cells.}
}
\value{
A tibble.
}
\description{
Create base grid for TICA
}
\examples{
\dontrun{
base_grid <- create_grid(resolution = 0.2, full_cells = TRUE)
}

}
