% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visualizations.R
\name{create_visualizations}
\alias{create_visualizations}
\alias{eda_histogram}
\alias{eda_cumulative_distribution}
\alias{eda_spatial_distribution}
\alias{eda_time_series}
\title{Title Create data visualizations}
\usage{
create_visualizations(
  f,
  data,
  group_variable,
  viz_title = NULL,
  x_title = NULL,
  y_title = NULL,
  out_format = c("png", "rdata"),
  out_filename = NULL,
  out_path = NULL,
  out_width = 15,
  out_height = 11,
  ...
)

eda_histogram(data, variable, n_bins, scale_transform, x_lim, ...)

eda_cumulative_distribution(
  data,
  variable,
  quantiles_list,
  scale_transform,
  ...
)

eda_spatial_distribution(
  data = NULL,
  variable = NULL,
  base_map = NULL,
  animated = FALSE
)

eda_time_series(
  data = NULL,
  variable = NULL,
  ts_type = c("step", "bar", "line")
)
}
\arguments{
\item{f}{Type of visualization}

\item{data}{A tibble to provide data}

\item{group_variable}{Set variable to create facet label}

\item{viz_title}{Visualization title}

\item{x_title}{Horizontal axis title}

\item{y_title}{Vertical axis title}

\item{out_format}{The format of the visualization output}

\item{out_filename}{Name of the file}

\item{out_path}{Path of the file}

\item{out_width}{Width in centimeters of the .png file}

\item{out_height}{Height in centimeters of the .png file}

\item{...}{Parameters passed to visualization functions}

\item{variable}{The variable to be represented in the visualization}

\item{scale_transform}{A transform for the horizontal axis}

\item{quantiles_list}{List of percentiles to be calculated}

\item{base_map}{An sf object to serve as base map}

\item{animated}{Should the visualization be animated?}

\item{ts_type}{The type of the time series visualization}
}
\value{
A ggplot
}
\description{
Title Create data visualizations
}
\examples{
\dontrun{
  create_visualizations()
}

}
