% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visualizations.R
\name{create_visualization}
\alias{create_visualization}
\alias{eda_histogram}
\alias{eda_cumulative_distribution}
\alias{eda_spatial_distribution}
\alias{eda_timeseries}
\alias{eda_colsum}
\alias{eda_summary_table}
\alias{eda_observations}
\alias{results_scatterplot}
\alias{results_mae_distribution}
\alias{results_mae_timeseries}
\alias{results_timeseries}
\alias{results_spatial}
\alias{results_mae_spatial}
\alias{method_samples}
\alias{explain_summary}
\alias{explain_dependence}
\alias{method_grid}
\alias{method_grid_area}
\title{Title Create data visualizations}
\usage{
create_visualization(
  f,
  group_variable = NA_character_,
  viz_title = NA_character_,
  x_title = NA_character_,
  y_title = NA_character_,
  out_filename = NA_character_,
  ...
)

eda_histogram(data, variable, ...)

eda_cumulative_distribution(
  data,
  variable,
  quantiles_list,
  scale_transform,
  ...
)

eda_spatial_distribution(data, variable, variable_label, summary, ...)

eda_timeseries(data, variable, ...)

eda_colsum(data, variable, cat_variable, scale_transform, ...)

eda_summary_table(data, table_title = parent.frame()$viz_title, ...)

eda_observations(data, variable = NA_character_, ...)

results_scatterplot(data, ...)

results_mae_distribution(data, ...)

results_mae_timeseries(data, ...)

results_timeseries(data, ...)

results_spatial(data, ...)

results_mae_spatial(data, ...)

method_samples(data, ...)

explain_summary(data, variable, ...)

explain_dependence(data, variable, ...)

method_grid(data, ...)

method_grid_area(data, ...)
}
\arguments{
\item{f}{Type of visualization}

\item{group_variable}{Set variable to create facet label}

\item{viz_title}{Visualization title}

\item{x_title}{Horizontal axis title}

\item{y_title}{Vertical axis title}

\item{out_filename}{Name of the file}

\item{...}{Parameters passed to visualization functions}

\item{data}{A tibble to provide data}

\item{variable}{The variable to be represented in the visualization}

\item{quantiles_list}{List of percentiles to be calculated}

\item{scale_transform}{A transform for the horizontal axis}

\item{variable_label}{Set the labels of the table variables}

\item{cat_variable}{Set the catgorical variable}

\item{out_format}{The format of the visualization output}

\item{n_bins}{Number of bins of histogram}

\item{x_lim}{Limit of x axis}

\item{rowname_variable}{Set the variable used as table row name}
}
\value{
A ggplot
}
\description{
Title Create data visualizations
}
\examples{
\dontrun{
  create_visualizations()
}

}
