% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visualizations.R
\name{create_visualizations}
\alias{create_visualizations}
\alias{eda_histogram}
\alias{eda_cumulative_distribution}
\alias{eda_spatial_distribution}
\alias{eda_time_series}
\title{Title Create data visualizations}
\usage{
create_visualizations()

eda_histogram(
  data = NULL,
  variable = NULL,
  n_bins = 30,
  scale_transform = "identity",
  xlim = NULL
)

eda_cumulative_distribution(
  data = NULL,
  variable = NULL,
  quantiles_list = c(0, 0.01, 0.05, 0.25, 0.5, 0.75, 0.95, 0.99, 1),
  scale_transform = "identity"
)

eda_spatial_distribution(
  data = NULL,
  variable = NULL,
  base_map = NULL,
  animated = FALSE
)

eda_time_series(
  data = NULL,
  variable = NULL,
  ts_type = c("step", "bar", "line")
)
}
\arguments{
\item{data}{A tibble to provide data}

\item{variable}{The variable to be represented in the visualization}

\item{n_bins}{Number of bins for the visualization}

\item{scale_transform}{A transform for the horizontal axis}

\item{xlim}{The limits for the horizontal axis}

\item{quantiles_list}{List of percentiles to be calculated}

\item{base_map}{An sf object to serve as base map}

\item{animated}{Should the visualization be animated?}

\item{ts_type}{The type of the time series visualization}
}
\value{
A ggplot
}
\description{
Title Create data visualizations
}
\examples{
\dontrun{
  eda_histogram()
}

}
