% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_external_data.R
\name{download_external_data}
\alias{download_external_data}
\alias{download_aoi}
\alias{download_land_use}
\alias{download_conservation_units}
\alias{download_indigenous_lands}
\alias{download_federal_roads}
\title{Download external data}
\usage{
download_external_data(
  dest_dir = "./data/external/",
  timespan = 1985:2021,
  crs,
  ...
)

download_aoi(dest_dir = "./data/external/", crs = "EPSG:5880")

download_land_use(
  dest_dir = "./data/external/",
  aoi_path = "./data/external/aoi/aoi.fgb",
  timespan = 1985:2021,
  crs = "EPSG:5880",
  agg_fact = 5
)

download_conservation_units(
  dest_dir = "./data/external/",
  crs = "EPSG:5880",
  ...
)

download_indigenous_lands(
  dest_dir = "./data/external/",
  crs = "EPSG:5880",
  ...
)

download_federal_roads(dest_dir = "./data/external/", crs = "EPSG:5880", ...)
}
\arguments{
\item{dest_dir}{Destination directory for external files.}

\item{timespan}{Set the years to download data.}

\item{crs}{Set the Coordinate Reference System to be used}

\item{aoi_path}{Set path to data with area of interest limits.}

\item{agg_fact}{Set the factor to aggregate raster cells}
}
\value{
A tibble
}
\description{
Download external data
}
