% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_external_data.R
\name{download_external_data}
\alias{download_external_data}
\alias{download_aoi}
\alias{download_land_use}
\alias{download_conservation_units}
\alias{download_indigenous_lands}
\title{Download external data}
\usage{
download_external_data(dest_dir = "./data/external/", timespan = 1985:2021)

download_aoi(dest_dir = "./data/external/")

download_land_use(
  dest_dir = "./data/external/",
  aoi_path = "./data/external/aoi/aoi.fgb",
  timespan = 1985:2021
)

download_conservation_units(dest_dir = "./data/external/")

download_indigenous_lands(dest_dir = "./data/external/")
}
\arguments{
\item{dest_dir}{Destination directory for external files.}

\item{timespan}{Set the years to download data.}

\item{aoi_path}{Set path to data with area of interest limits.}
}
\value{
A tibble
}
\description{
Download external data
}
