% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_external_data.R
\name{process_external_data}
\alias{process_external_data}
\title{Process external data to base grid}
\usage{
process_external_data(data_name, read_function, process_function)
}
\arguments{
\item{external_data}{a}

\item{preprocess_function}{a}
}
\value{
An sf object
}
\description{
Process external data to base grid
}
