% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_external_data.R
\name{process_external_data}
\alias{process_external_data}
\alias{process_conservation_units}
\alias{process_indigenous_lands}
\title{Process external data to base grid}
\usage{
process_external_data()

process_conservation_units(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/uc/uc.fgb",
  timespan = 1985:2021
)

process_indigenous_lands(
  base_grid_path = "./data/base_grid.fgb",
  external_data_path = "./data/external/il/il.fgb",
  timespan = 1985:2021
)
}
\arguments{
\item{base_grid_path}{Path to the base grid file}

\item{external_data_path}{Set path to external data file}

\item{timespan}{Set the years for the data in the base grid}
}
\value{
A tibble
}
\description{
Process external data to base grid
}
\examples{
\dontrun{
  uc_grid <- process_conservation_units("./inst/base_grid.fgb", 1985:2021)
}

}
